fix:coundown folder & move to home
