+++ src/app/routes/index.tsx
@@ -12,7 +12,6 @@ import { createBrowserRouter } from 'react-router-dom';
 import UserProfile from '../features/developers/components/user-profile';
 import PortfolioFormview from '../features/portfolio/components/portfolio-form-view';
 import { authRoutes } from './auth';
-import Countdown from '../features/countdown/countdown';
 
 const router = createBrowserRouter([
   {
@@ -33,8 +32,6 @@ const router = createBrowserRouter([
       { path: '/approved-ideas', element: <ApprovedIdeaPage /> },
       { path: '/ideas', element: <IdeaPage /> },
       { path: 'profile/:username', element: <UserProfile /> },
-
-      {path: '/count-down', element: <Countdown/>}
     ],
   },
